+++ src/main/java/lotto/shop/Money.java
